+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 -   Node: Allow routing Cluster requests by address. ([#1021](https://github.com/aws/glide-for-redis/pull/1021))
 * Python: Added HSETNX command. ([#954](https://github.com/aws/glide-for-redis/pull/954))
 * Python: Added SISMEMBER command ([#971](https://github.com/aws/glide-for-redis/pull/971))
+* Python: Added ZRANGE command ([#906](https://github.com/aws/glide-for-redis/pull/906))
 
 ## 0.2.0 (2024-02-11)